third_party/WebKit/LayoutTests/external/wpt/html/dom/self-origin.sub-expected.txt

This is a testharness.js-based test.
PASS Assigning blob url 
PASS Should have the right origin for cross-origin subframe 
PASS Should have the right origin for cross-origin subframe after setting document.domain 
FAIL Should have the right origin for IDN subframe assert_equals: expected "http://élève.web-platform.test:8081" but got "http://xn--lve-6lad.web-platform.test:8081"
FAIL Should have the right origin for IDN subframe after setting document.domain assert_equals: expected "http://élève.web-platform.test:8081" but got "http://xn--lve-6lad.web-platform.test:8081"
PASS Should have the right origin for sandboxed iframe 
PASS We should have the right origin for our page 
PASS about:blank subframe origins 
PASS blob: subframe origins 
PASS javascript: subframe origins 
PASS srcdoc subframe origins 
Harness: the test ran to completion.
